requirements.txt
Flask[async]==2.3.3
python-telegram-bot==20.3
gunicorn==21.2.0
requests==2.31.0